Add files via upload
TOver the years, the exploitation and pollution of natural resources have created the need for renewable and environment-friendly products. One of these products is Electric Vehicles. Electric Vehicles are the replacement for petroleum-based vehicles. They are one of the emerging technologies as well as eco-friendly and viable. The replacement of internal combustion engines with electric enginesMany countries around the globe have implemented this technology and are contributing towards amelioration of the environment. We are going to see the opportunities and challenged faced in India over implementing electric vehicles.